+++ src/MonitorServiceProvider.php
@@ -6,7 +6,6 @@ use Drcantagalo\LaravelMonitor\Console\Commands\MonitorAuditPathsCommand;
 use Drcantagalo\LaravelMonitor\Console\Commands\MonitorExportDenylistCommand;
 use Drcantagalo\LaravelMonitor\Console\Commands\MonitorInstallCommand;
 use Drcantagalo\LaravelMonitor\Console\Commands\MonitorPruneCommand;
-use Drcantagalo\LaravelMonitor\Console\Commands\MonitorRecalculateVisitsCommand;
 use Drcantagalo\LaravelMonitor\Console\Commands\MonitorUpdateCommand;
 use Drcantagalo\LaravelMonitor\Http\Middleware\MonitorMethod;
 use Drcantagalo\LaravelMonitor\Support\Monitor;
@@ -74,7 +73,6 @@ class MonitorServiceProvider extends ServiceProvider
                 MonitorUpdateCommand::class,
                 MonitorAuditPathsCommand::class,
                 MonitorPruneCommand::class,
-                MonitorRecalculateVisitsCommand::class,
             ]);
         }